requirements.txt
awscli==1.6.2
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
bcdoc==0.12.2
beautifulsoup4==4.5.1
boto==2.43.0
boto3==1.4.1
botocore==1.4.78
bz2file==0.98
CacheControl==0.11.7
certifi==2016.9.26
chainer==1.19.0
cligj==0.4.0
colorama==0.2.5
configparser==3.5.0
cycler==0.10.0
decorator==4.0.10
docutils==0.12
entrypoints==0.2.2
enum34==1.1.6
fastdtw==0.3.0
filechunkio==1.8
filelock==2.0.7
functools32==3.2.3.post2
futures==3.0.5
gensim==0.13.3
genson==0.1.0
geopy==1.11.0
google-api-python-client==1.5.4
httplib2==0.9.2
ipykernel==4.5.0
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
iso3166==0.7
Jinja2==2.8
jmespath==0.9.0
joblib==0.10.3
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.0
lda==1.0.4
mapbox==0.10.1
MarkupSafe==0.23
matplotlib==2.0
mistune==0.7.3
mpmath==0.19
mrjob==0.5.6
multipledispatch==0.4.9
nbconvert==4.2.0
nbformat==4.1.0
networkx==1.11
nltk==3.2.1
nose==1.3.7
notebook==4.2.3
oauth2==1.9.0.post1
oauth2client==4.0.0
oauthlib==2.0.0
pathlib2==2.1.0
patsy==0.4.1
pbr==1.10.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.4.2
polyline==1.3.1
powerlaw==1.4.1
prettytable==0.7.2
prompt-toolkit==1.0.7
protobuf==3.1.0.post1
psclient==2.0
ptyprocess==0.5.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pyclustering==0.6.6
Pygments==2.1.3
pyparsing==2.1.9
pyproj==1.9.5.1
pyshp==1.2.10
python-dateutil==2.6.0
pytz==2016.6.1
PyYAML==3.12
pyzmq==15.4.0
qtconsole==4.2.1
requests==2.11.1
requests-oauthlib==0.7.0
rsa==3.4.2
Rtree==0.8.3
s3transfer==0.1.9
seaborn==0.7.1
selenium==2.53.6
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
smart-open==1.3.5
sseclient==0.0.8
statsmodels==0.6.1
tabulate==0.7.7
termcolor==1.1.0
terminado==0.6
tornado==4.4.1
traitlets==4.3.0
tweepy==3.5.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.18
wcwidth==0.1.7
widgetsnbextension==1.2.6